requirements.txt
fastapi
uvicorn
line-bot-sdk
google-generativeai
python-dotenv
requests
pillow
mangum